+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JW1VG6MWGHWARMC4ZZJKBJ3C"
+        "01JW5Q0BZ42NKVB22RAWMK43H3"
     )
     r"""The id of the resource"""